requirements.txt
bcrypt==4.3.0
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
greenlet==3.2.4
gunicorn==20.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
python-dotenv==0.21.0
SQLAlchemy==1.4.46
typing_extensions==4.14.1
Werkzeug==2.3.2
WTForms==3.2.1
Flask-Migrate
stripe
Flask-Mail
APScheduler
twilio